refactor(stackspot): extract shared logic to eliminate stacking-lottery duplication
Moves POT_DEPLOYER, PLATFORM_ADDRESS, PLATFORM_CONTRACT, KNOWN_POTS,
parseContractName, and callPotReadOnly into src/lib/utils/stackspot-shared.ts.
Both stackspot.ts and stacking-lottery.ts become thin CLI wrappers (~60 lines each)
that import all business logic from the shared module.

Bug fixes and pot list updates now only need to be made in one place.

Closes #189
