+++ stackspot/stackspot.ts
@@ -10,119 +10,27 @@
 import { Command } from "commander";
 import { NETWORK, getExplorerTxUrl } from "../src/lib/config/networks.js";
 import { getAccount } from "../src/lib/services/x402.service.js";
-import { getHiroApi } from "../src/lib/services/hiro-api.js";
 import { callContract } from "../src/lib/transactions/builder.js";
 import { printJson, handleError } from "../src/lib/utils/cli.js";
 import {
   uintCV,
   contractPrincipalCV,
   PostConditionMode,
-  type ClarityValue,
-  deserializeCV,
-  cvToJSON,
 } from "@stacks/transactions";
+import {
+  PLATFORM_ADDRESS,
+  PLATFORM_CONTRACT,
+  KNOWN_POTS,
+  parseContractName,
+  callPotReadOnly,
+} from "../src/lib/utils/stackspot-shared.js";
 
-// ---------------------------------------------------------------------------
-// Constants
-// ---------------------------------------------------------------------------
-
-const POT_DEPLOYER = "SPT4SQP5RC1BFAJEQKBHZMXQ8NQ7G118F335BD85";
-const PLATFORM_ADDRESS = "SP7FSE31MWSJJFTQBEQ1TT6TF3G4J6GDKE81SWD9";
-const PLATFORM_CONTRACT = "stackspots";
-
-interface PotInfo {
-  name: string;
-  contractName: string;
-  maxParticipants: number;
-  minAmountStx: number;
-  deployer: string;
-}
-
-const KNOWN_POTS: PotInfo[] = [
-  {
-    name: "Genesis",
-    contractName: "Genesis",
-    maxParticipants: 2,
-    minAmountStx: 20,
-    deployer: POT_DEPLOYER,
-  },
-  {
-    name: "BuildOnBitcoin",
-    contractName: "BuildOnBitcoin",
-    maxParticipants: 10,
-    minAmountStx: 100,
-    deployer: POT_DEPLOYER,
-  },
-  {
-    name: "STXLFG",
-    contractName: "STXLFG",
-    maxParticipants: 100,
-    minAmountStx: 21,
-    deployer: POT_DEPLOYER,
-  },
-];
-
-// ---------------------------------------------------------------------------
-// Helpers
-// ---------------------------------------------------------------------------
-
-/**
- * Parse a contract name that may be fully qualified (deployer.name) or bare (name).
- * Returns { deployer, contractName }.
- */
-function parseContractName(input: string): {
-  deployer: string;
-  contractName: string;
-} {
-  if (input.includes(".")) {
-    const [deployer, ...rest] = input.split(".");
-    return { deployer, contractName: rest.join(".") };
-  }
-  return { deployer: POT_DEPLOYER, contractName: input };
-}
-
-/**
- * Call a read-only function on a pot contract.
- * Accepts either a bare contract name or fully qualified deployer.name.
- * Returns the deserialized Clarity value as a JSON-friendly object.
- */
-async function callPotReadOnly(
-  contractNameOrId: string,
-  functionName: string,
-  args: ClarityValue[]
-): Promise<unknown> {
-  const hiro = getHiroApi(NETWORK);
-  const { deployer, contractName } = parseContractName(contractNameOrId);
-  const contractId = `${deployer}.${contractName}`;
-  const result = await hiro.callReadOnlyFunction(
-    contractId,
-    functionName,
-    args,
-    deployer
-  );
-  if (!result.okay) {
-    throw new Error(
-      `Read-only call ${functionName} failed: ${result.cause ?? "unknown error"}`
-    );
-  }
-  if (!result.result) {
-    return null;
-  }
-  const hex = result.result.startsWith("0x")
-    ? result.result.slice(2)
-    : result.result;
-  const cv = deserializeCV(Buffer.from(hex, "hex"));
-  return cvToJSON(cv);
-}
-
-// ---------------------------------------------------------------------------
-// Program
-// ---------------------------------------------------------------------------
+const SKILL_NAME = "stackspot";
 
 const program = new Command();
 
 program
-  .name("stackspot")
+  .name(SKILL_NAME)
   .description(
     "Stacking lottery pots on stackspot.app — pool STX into pots that stack via PoX, " +
       "VRF picks a random winner for sBTC rewards, all participants get their STX back. Mainnet-only."
@@ -142,7 +50,7 @@ program
     try {
       if (NETWORK !== "mainnet") {
         throw new Error(
-          "stackspot skill is mainnet-only. Set NETWORK=mainnet to use this skill."
+          `${SKILL_NAME} skill is mainnet-only. Set NETWORK=mainnet to use this skill.`
         );
       }
 
@@ -202,7 +110,7 @@ program
     try {
       if (NETWORK !== "mainnet") {
         throw new Error(
-          "stackspot skill is mainnet-only. Set NETWORK=mainnet to use this skill."
+          `${SKILL_NAME} skill is mainnet-only. Set NETWORK=mainnet to use this skill.`
         );
       }
 
@@ -257,7 +165,7 @@ program
     try {
       if (NETWORK !== "mainnet") {
         throw new Error(
-          "stackspot skill is mainnet-only. Set NETWORK=mainnet to use this skill."
+          `${SKILL_NAME} skill is mainnet-only. Set NETWORK=mainnet to use this skill.`
         );
       }
 
@@ -268,7 +176,6 @@ program
 
       const parsed = parseContractName(opts.contractName);
 
-      // Warn if amount is below the known minimum for this pot
       const knownPot = KNOWN_POTS.find(
         (p) => p.contractName === parsed.contractName
       );
@@ -325,7 +232,7 @@ program
     try {
       if (NETWORK !== "mainnet") {
         throw new Error(
-          "stackspot skill is mainnet-only. Set NETWORK=mainnet to use this skill."
+          `${SKILL_NAME} skill is mainnet-only. Set NETWORK=mainnet to use this skill.`
         );
       }
 
@@ -372,7 +279,7 @@ program
     try {
       if (NETWORK !== "mainnet") {
         throw new Error(
-          "stackspot skill is mainnet-only. Set NETWORK=mainnet to use this skill."
+          `${SKILL_NAME} skill is mainnet-only. Set NETWORK=mainnet to use this skill.`
         );
       }
 
@@ -419,7 +326,7 @@ program
     try {
       if (NETWORK !== "mainnet") {
         throw new Error(
-          "stackspot skill is mainnet-only. Set NETWORK=mainnet to use this skill."
+          `${SKILL_NAME} skill is mainnet-only. Set NETWORK=mainnet to use this skill.`
         );
       }
